requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
httpx==0.24.1
beautifulsoup4==4.12.2
pdfplumber==0.7.6
pandas==2.2.2
pillow==10.0.0
pytesseract==0.3.10
python-multipart==0.0.9

# Optional (only if local whisper is used)
# faster-whisper==1.0.0
